+++ src/NosCore.Packets/ServerPackets/CModePacket.cs
@@ -34,13 +34,13 @@ namespace NosCore.Packets.ServerPackets
         public long VisualId { get; set; }
 
         [PacketIndex(2)]
-        public byte Morph { get; set; }
+        public short Morph { get; set; }
 
         [PacketIndex(3)]
         public byte MorphUpgrade { get; set; }
 
         [PacketIndex(4)]
-        public byte MorphDesign { get; set; }
+        public short MorphDesign { get; set; }
 
         [PacketIndex(5, IsOptional = true)]
         public byte MorphBonus { get; set; }
